added timeouts to infinite loops in DHT11.cpp
converted ERROR_TIMEOUT in DHT11.h from int to uint8_t to resolve return type issues

+++ onboard/src/offboard_comms/sketches/peripheral/DHT11.h
@@ -65,7 +65,7 @@ public:
 
   // Constants to represent error codes.
   static const int ERROR_CHECKSUM = 254;    // Error code indicating checksum mismatch.
-  static const int ERROR_TIMEOUT = 253;     // Error code indicating a timeout occurred during reading.
+  static const uint8_t ERROR_TIMEOUT = 253;     // Error code indicating a timeout occurred during reading.
   static const int TIMEOUT_DURATION = 10; // Duration (in milliseconds) to wait before timing out. This value has been modified from the original 1000 milliseconds to 10 milliseconds.
 
   /**
